+++ ChangeLog.md
@@ -18,11 +18,15 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 
 1. Ported rule `hopperCountersUnlimitedSpeed`.
 2. Ported rule `hopperNoItemCost`.
+3. Fix a logger format bug about which causes game crash.
+4. Counter result is displayed with more time units (gt, s, min, h, d, week).
 
 ## Lifetime Tracker
 
 1. Add rule `lifetimeTrackBySize`.
 2. Chicken jockeys and spider jockeys will be tracked.
+3. The tracker reports standard error, standard deviation, and variance, of the tracked lifetimes.
+4. The tracker reports average mobcap payload.
 
 ## Miscellaneous